+++ src/CSET/cset_workflow/app/fetch_fcst/bin/fetch_data.py
@@ -1,4 +1,16 @@
-#! /usr/bin/env python3
+# © Crown copyright, Met Office (2022-2025) and CSET contributors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 """Retrieve the files from the filesystem for the current cycle point."""
 
@@ -207,7 +219,7 @@ def _template_file_path(
     return sorted(paths)
 
 
-def fetch_data(file_retriever: FileRetrieverABC = FilesystemFileRetriever):
+def fetch_data(file_retriever: FileRetrieverABC):
     """Fetch the data for a model.
 
     The following environment variables need to be set:
@@ -223,7 +235,7 @@ def fetch_data(file_retriever: FileRetrieverABC = FilesystemFileRetriever):
     Parameters
     ----------
     file_retriever: FileRetriever
-        FileRetriever implementation to use. Defaults to FilesystemFileRetriever.
+        FileRetriever implementation to use.
 
     Raises
     ------
