--- src/CSET/cset_workflow/app/finish_website/bin/finish-website.py
@@ -1,7 +1,0 @@
-#! /usr/bin/env python3
-
-"""Write finished status to website front page."""
-
-import CSET._workflow_utils.finish_website
-
-CSET._workflow_utils.finish_website.run()
