+++ src/CSET/cset_workflow/app/finish_website/bin/finish_website.py
@@ -1,6 +1,20 @@
 #! /usr/bin/env python3
+# © Crown copyright, Met Office (2022-2025) and CSET contributors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
-"""Write finished status to website front page.
+"""
+Write finished status to website front page.
 
 Constructs the plot index, and does the final update to the workflow status on
 the front page of the web interface.
