+++ src/CSET/cset_workflow/app/finish_website/rose-app.conf
@@ -1,2 +1,2 @@
 [command]
-default=app_env_wrapper finish-website.py
+default=app_env_wrapper finish_website.py
