--- src/CSET/cset_workflow/app/run_cset_recipe/bin/run-cset-recipe.py
@@ -1,7 +1,0 @@
-#! /usr/bin/env python3
-
-"""Run a recipe with the CSET CLI."""
-
-import CSET._workflow_utils.run_cset_recipe
-
-CSET._workflow_utils.run_cset_recipe.run()
