+++ src/CSET/cset_workflow/app/run_cset_recipe/bin/run_cset_recipe.py
@@ -1,4 +1,17 @@
 #! /usr/bin/env python3
+# © Crown copyright, Met Office (2022-2025) and CSET contributors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 """Run a recipe with the CSET CLI."""
 
@@ -65,3 +78,7 @@ def run():
         run_recipe_steps()
     except subprocess.CalledProcessError:
         sys.exit(1)
+
+
+if __name__ == "__main__":
+    run()
