+++ src/CSET/cset_workflow/app/run_cset_recipe/rose-app.conf
@@ -1,2 +1,2 @@
 [command]
-default=app_env_wrapper run-cset-recipe.py
+default=app_env_wrapper run_cset_recipe.py
