--- src/CSET/cset_workflow/app/send_email/bin/send-email.py
@@ -1,7 +1,0 @@
-#! /usr/bin/env python3
-
-"""Send a notification email linking to the output page."""
-
-import CSET._workflow_utils.send_email
-
-CSET._workflow_utils.send_email.run()
