+++ src/CSET/cset_workflow/app/send_email/bin/send_email.py
@@ -1,6 +1,20 @@
 #! /usr/bin/env python3
+# © Crown copyright, Met Office (2022-2025) and CSET contributors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
-"""Send a notification email linking to the output page.
+"""
+Send a notification email linking to the output page.
 
 Sends a notification email to the user of the workflow when it has completed,
 with a link to the output webpage.
@@ -23,7 +37,7 @@ def get_home_page_address():
 
     path = web_dir.split("/public_html/")
     if len(path) == 1:
-        path = os.getenv("WEB_DIR").split("/Public/")
+        path = web_dir.split("/Public/")
     if len(path) == 1:
         raise ValueError("Cannot determine web address.")
     base_address = web_addr.strip("/")
@@ -42,3 +56,7 @@ def run():
         check=True,
         shell=True,
     )
+
+
+if __name__ == "__main__":
+    run()
