+++ src/CSET/cset_workflow/app/send_email/rose-app.conf
@@ -1,2 +1,2 @@
 [command]
-default=app_env_wrapper send-email.py
+default=app_env_wrapper send_email.py
