+++ tests/workflow_utils/test_finish_website.py
@@ -18,7 +18,7 @@ import json
 import logging
 import re
 
-from CSET._workflow_utils import finish_website
+from CSET.cset_workflow.app.finish_website.bin import finish_website
 
 
 def test_copy_rose_config(monkeypatch, tmp_path):
